some minor refactoring, made serial backend work with new refactored tasking
